+++ src/main/java/carracing/view/OutputView.java
@@ -1,4 +1,4 @@
-package carracing.ui;
+package carracing.view;
 
 import java.util.List;
